+++ src/tables.js
@@ -64,13 +64,13 @@
 
 // Global settings /////////////////////////////////////////////////////
 
-/*global modules, Point, Morph, fontHeight, SliderMorph, isString, detect,
-MorphicPreferences, FrameMorph, HandleMorph, DialogBoxMorph, StringMorph,
-SpriteMorph, Context, Costume, BlockEditorMorph, SymbolMorph, List, IDE_Morph,
-SyntaxElementMorph, MenuMorph, SpriteBubbleMorph, SpeechBubbleMorph, Sound,
-CellMorph, ListWatcherMorph, isNil, BoxMorph, Variable, isSnapObject*/
+/*global modules, Point, Morph, fontHeight, SliderMorph, isString, detect, List,
+MorphicPreferences, FrameMorph, HandleMorph, DialogBoxMorph, StringMorph, isNil,
+SpriteMorph, Context, Costume, BlockEditorMorph, SymbolMorph, IDE_Morph, Sound,
+SyntaxElementMorph, MenuMorph, SpriteBubbleMorph, SpeechBubbleMorph, CellMorph,
+ListWatcherMorph, BoxMorph, Variable, isSnapObject, useBlurredShadows*/
 
-modules.tables = '2020-May-18';
+modules.tables = '2020-October-06';
 
 var Table;
 var TableCellMorph;
@@ -358,10 +358,12 @@ TableCellMorph.prototype.render = function (ctx) {
     if (dta instanceof HTMLCanvasElement) {
         x = Math.max((width - dta.width) / 2, 0);
         y = Math.max((height - dta.height) / 2, 0);
-        ctx.shadowOffsetX = 4;
-        ctx.shadowOffsetY = 4;
-        ctx.shadowBlur = 4;
-        ctx.shadowColor = 'lightgray';
+        if (useBlurredShadows) {
+            ctx.shadowOffsetX = 4;
+            ctx.shadowOffsetY = 4;
+            ctx.shadowBlur = 4;
+            ctx.shadowColor = 'lightgray';
+        }
         ctx.drawImage(dta, x, y);
     } else { // text
         ctx.font = font;